fix(database): close prepared statement in defer
Before this commit, prepared statement's Close() method was called only
if the happy path was taken.

This commit makes sure that regardless of the outcome of the subsequent
database query, the prepared statement is Closed().

Closing prepared statement was a source of great success in api-server
with regards to memory usage.

We should most probably refactor the prepared statement usage across our
services, because we're supposed to create them once and re-use them
multiple times.